refactor(dbus): partition windowtrackingadaptor and enginewiring by concern
